thang2 merge 29/9

+++ AtpMan/src/java/DAO/InvoiceDAO.java
@@ -76,7 +76,7 @@ public class InvoiceDAO {
                 service.setName(rs.getString("name"));
                 service.setType(rs.getString("type"));
                 service.setDescription(rs.getString("description"));
-                service.setFee(rs.getDouble("fee"));
+                service.setFee(rs.getBigDecimal("fee"));
 
                 // declare serviceContract object
                 ServiceContract serviceContract = new ServiceContract();
